Refactor: url 및 컴포넌트명 변경 및 폴더 구조 변경, 팀 엔트리 분리
