+++ src/app/components/common/background/index.tsx
