+++ src/app/components/common/background/styles.ts
