+++ src/app/components/common/loading/styles.ts
