+++ src/app/components/common/not-found/styles.ts
