+++ src/app/components/footer/styles.ts
