+++ src/app/components/selection/teamSelection/styles.ts
