+++ src/app/entry/[entryId]/components/entryItem/styles.ts
