+++ src/app/entry/[entryId]/components/infoBox/index.tsx
