+++ src/app/entry/[entryId]/components/infoBox/styles.ts
