+++ src/app/entry/[entryId]/components/lineUpInfo/buffItem/index.tsx
