+++ src/app/entry/[entryId]/components/lineUpInfo/index.tsx
