+++ src/app/entry/[entryId]/components/lineUpInfo/styles.ts
