+++ src/app/entry/[entryId]/components/lineUpInfo/total/index.tsx
