+++ src/app/entry/[entryId]/components/lineUpInfo/total/styles.ts
