+++ src/app/entry/[entryId]/components/playerDetail/index.tsx
