+++ src/app/entry/[entryId]/components/playerDetail/styles.ts
