+++ src/app/entry/[entryId]/components/playerSimpleInfo/index.tsx
