+++ src/app/entry/[entryId]/components/playerSimpleInfo/stat/index.tsx
