+++ src/app/entry/[entryId]/components/playerSimpleInfo/stat/styles.ts
