+++ src/app/entry/[entryId]/components/playerSimpleInfo/styles.ts
