+++ src/app/entry/[entryId]/components/teamEntry/styles.ts
