+++ src/app/entry/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from 'next/navigation';
 
-export default function Lineup() {
+export default function Page() {
   redirect('/');
 }
